+++ app/Http/Controllers/DocumentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Document;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\DocumentRequest;
 use App\Http\Services\DocumentService;
 
@@ -25,6 +26,10 @@ class DocumentController extends Controller
 
     public function destroy(DocumentService $service, Document $document)
     {
+        if (!Gate::allows('delete-document')) {
+            abort(403);
+        }
+
         $caselaw = $document->caselaw;
         $service->destroy($document);
         return redirect()->route('caselaw.document.index', $document->caselaw)
@@ -33,6 +38,10 @@ class DocumentController extends Controller
 
     public function download(DocumentService $service, Document $document)
     {
+        if (!Gate::allows('view-document')) {
+            abort(403);
+        }
+
         return $service->download($document);
     }
 }
