+++ app/Http/Controllers/InvoiceController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Invoice;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\InvoiceRequest;
 use App\Http\Services\InvoiceService;
 
@@ -11,12 +12,9 @@ class InvoiceController extends Controller
 {
     public function index()
     {
-        //
-    }
-
-    public function create()
-    {
-        //
+        if (!Gate::allows('view-invoice')) {
+            abort(403);
+        }
     }
 
     public function store(InvoiceRequest $request, InvoiceService $service)
@@ -24,16 +22,6 @@ class InvoiceController extends Controller
         $invoice = $service->store($request);
         return redirect($request->has('redirect') ? $request->input('redirect') : '/invoice/' . $invoice->id)
         ->with('status', 'Berhasil membuat invoice baru');
-    }
-
-    public function show(Invoice $invoice)
-    {
-        //
-    }
-
-    public function edit(Invoice $invoice)
-    {
-        //
     }
 
     public function update(InvoiceRequest $request, InvoiceService $service, Invoice $invoice)
@@ -45,6 +33,10 @@ class InvoiceController extends Controller
 
     public function destroy(InvoiceService $service, Invoice $invoice)
     {
+        if (!Gate::allows('delete-invoice')) {
+            abort(403);
+        }
+
         $caselaw = $invoice->caselaw;
         $service->destroy($invoice);
         return redirect()->route('caselaw.invoice.index', $caselaw)
