+++ app/Http/Controllers/MessageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Caselaw;
 use App\Models\Message;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\MessageRequest;
 use App\Http\Services\MessageService;
 
@@ -19,6 +20,10 @@ class MessageController extends Controller
 
     public function destroy(MessageService $service, Message $message)
     {
+        if (!Gate::allows('delete-message')) {
+            abort(403);
+        }
+        
         $caselaw = $message->caselaw;
         $service->destroy($message);
         return redirect()->route('caselaw.message.index', $caselaw)
